+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Requires PHP `8.2`
 - Requires `innmind/time-continuum:~4.1`
 - Requires `innmind/immutable:~5.12`
+- `Innmind\TimeWarp\Halt` return an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`